pyproject.toml
[project]
name = "verifai"
version = "2.1.2"
description = "A toolkit for the formal design and analysis of systems that include artificial intelligence (AI) and machine learning (ML) components."
authors = [
	{ name = "Tommaso Dreossi" },
	{ name = "Daniel Fremont", email = "dfremont@ucsc.edu" },
	{ name = "Shromona Ghosh" },
	{ name = "Edward Kim", email = "ek65@berkeley.edu" },
	{ name = "Hadi Ravanbakhsh" },
	{ name = "Marcell Vazquez-Chanlatte" },
	{ name = "Sanjit A. Seshia" }
]
maintainers = [
	{ name = "Daniel Fremont", email = "dfremont@ucsc.edu" },
	{ name = "Edward Kim", email = "ek65@berkeley.edu" }
]
readme = "README.md"
requires-python = ">=3.8,<4.0"
license = "BSD-3-Clause"
classifiers = [
	"Operating System :: OS Independent",
	"Development Status :: 5 - Production/Stable",
	"Programming Language :: Python :: 3"
]

dependencies = [
	"numpy >= 1.21",
	"scipy ~= 1.7",
	"dotmap ~= 1.3",
	"metric-temporal-logic >= 0.1.5",
	"matplotlib ~= 3.2",
	"easydict ~= 1.9",
	"joblib ~= 1.2",
	"dill >= 0.3.1",
	"future >= 0.18.3",
	"pandas >= 1.3, <3",
	"scikit-learn ~= 1.0",
	"pygame ~= 2.1",
	"kmodes <= 0.10.2",
	"scenic >= 2.1.0b1, <4",
	"progressbar2 ~= 3.53",
	"networkx >= 2.6",
	"statsmodels >= 0.13.2"
]

[project.optional-dependencies]
bayesopt = [
	"GPy ~= 1.9",
	"GPyOpt ~= 1.2"
]
parallel = [
	"ray ~= 1.10"
]
examples = [
	"tensorflow ~= 2.8",
	"gym >= 0.22",
	"pyglet ~= 1.5",
	"opencv-python ~= 4.5",
	"pillow >= 9.3,<10.0",
	'pyproj ~= 3.0; python_version < "3.10"',
	'pyproj ~= 3.3; python_version >= "3.10"'
]
test = [ # minimum dependencies for running tests (used for tox virtualenvs)
	"pytest >= 6.2.5, <8",
	"pytest-randomly >= 3.2"
]
dev = [ # all dependencies from 'test' extra above
	"tox ~= 3.14",
	"sphinx >= 4.1.0, <6",
	"sphinx-rtd-theme >= 0.5.2",
	"sphinx-tabs ~= 3.4.1",
	"recommonmark >= 0.6.0"
]

[project.urls]
Repository = "https://github.com/BerkeleyLearnVerify/VerifAI"
Documentation = "https://verifai.readthedocs.io"

[build-system]
requires = ["flit_core >= 3.2, <4"]
build-backend = "flit_core.buildapi"